refactor : Jwt, Oauth Properties 패키지 구조 변경 (#559)
