+++ src/main/java/com/clubber/ClubberServer/domain/auth/component/UserWithdraw.java
@@ -4,7 +4,7 @@ import com.clubber.ClubberServer.domain.auth.dto.UnlinkKaKaoTarget;
 import com.clubber.ClubberServer.domain.auth.service.AuthService;
 import com.clubber.ClubberServer.domain.user.domain.User;
 import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.kakao.KakaoInfoClient;
-import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.properties.KakaoProperties;
+import com.clubber.ClubberServer.global.properties.KakaoProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
