+++ src/main/java/com/clubber/ClubberServer/global/config/ConfigurationPropertiesConfig.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.global.config;
 
 
-import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.properties.KakaoProperties;
+import com.clubber.ClubberServer.global.properties.KakaoProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
